Rename: Rename and move deposit validator for common use
